Renamed roslaunch to launch
Kept ros1/launch/roslaunch because it really is about roslaunch
