+++ setup.cfg
@@ -55,20 +55,20 @@ per-file-ignores =
   src/roswire/parameters/__init__.py:F401
   src/roswire/proxy/__init__.py:F401
   src/roswire/common/__init__.py:F401
-  src/roswire/common/roslaunch/__init__.py:F401
-  src/roswire/common/roslaunch/config/__init__.py:F401
-  src/roswire/common/roslaunch/config/env.py:N817
-  src/roswire/common/roslaunch/config/node.py:N817
-  src/roswire/common/roslaunch/config/launch.py:N817
-  src/roswire/common/roslaunch/config/parameter.py:N817
+  src/roswire/common/launch/__init__.py:F401
+  src/roswire/common/launch/config/__init__.py:F401
+  src/roswire/common/launch/config/env.py:N817
+  src/roswire/common/launch/config/node.py:N817
+  src/roswire/common/launch/config/launch.py:N817
+  src/roswire/common/launch/config/parameter.py:N817
   src/roswire/ros1/__init__.py:F401
   src/roswire/ros1/bag/__init__.py:F401
-  src/roswire/ros1/roslaunch/__init__.py:F401
-  src/roswire/ros1/roslaunch/reader.py:F811,E704,E501,N817
-  src/roswire/ros1/roslaunch/roslaunch.py:N817
+  src/roswire/ros1/launch/__init__.py:F401
+  src/roswire/ros1/launch/reader.py:F811,E704,E501,N817
+  src/roswire/ros1/launch/roslaunch.py:N817
   src/roswire/ros1/service/__init__.py:F401
   src/roswire/ros2/__init__.py:F401
-  src/roswire/ros2/roslaunch/__init__.py:F401
+  src/roswire/ros2/launch/__init__.py:F401
 
 [tox]
 envlist = py36, py37, py38
