+++ src/roswire/common/__init__.py
@@ -2,11 +2,11 @@
 from .action import ActionFormat
 from .base import Duration, Time
 from .format import FormatDatabase
+from .launch import ROSLaunchManager
 from .msg import Constant, Field, Message, MsgFormat
 from .node import Node
 from .node_manager import NodeManager
 from .package import Package, PackageDatabase
-from .roslaunch import ROSLaunchManager
 from .srv import SrvFormat
 from .state import SystemState
 from .type_db import TypeDatabase
