+++ src/roswire/common/launch/config/node.py
