--- src/roswire/common/roslaunch/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from .roslaunch import ROSLaunchManager
