+++ src/roswire/ros1/__init__.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
+from .launch import ROS1LaunchManager
 from .node import ROS1Node
 from .node_manager import ROS1NodeManager
 from .parameter import ParameterServer
 from .ros1 import ROS1
-from .roslaunch import ROS1LaunchManager
 from .service import Service, ServiceManager
 from .state import SystemStateProbe
