+++ src/roswire/ros2/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
+from .launch.launch import ROS2LaunchManager
 from .node_manager import ROS2NodeManager
 from .ros2 import ROS2
-from .roslaunch.launch import ROS2LaunchManager
 from .service_manager import ROS2ServiceManager
