+++ src/roswire/ros2/ros2.py
@@ -5,8 +5,8 @@ import typing
 
 import attr
 
+from .launch.launch import ROS2LaunchManager
 from .node_manager import ROS2NodeManager
-from .roslaunch.launch import ROS2LaunchManager
 from .service_manager import ROS2ServiceManager
 from .state import ROS2StateProbe
 from ..common import SystemState
